♻️ [Refactoring]: #489 centralize subpath closing for h and paint variants
- add immutable CurrentPath.closeSubpath
- add closeSubpathContext for shared close semantics
- refactor hHandler to use the shared context helper
